pyproject.toml
[project]
name = "planes-scraping"
version = "0.1.0"
description = "A web scraping project to extract data from the private-radar website."
authors = [
    {name = "Erick Cruz", email = "crzerick6@gmail.com"},
]
dependencies = [
    "attrs==24.2.0",
    "beautifulsoup4==4.12.3",
    "blinker==1.8.2",
    "certifi==2024.8.30",
    "click==8.1.7",
    "et-xmlfile==1.1.0",
    "flask==3.0.3",
    "Flask-Login==0.6.3",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Migrate>=4.0.0",
    "greenlet==3.1.1",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "idna==3.8",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.4",
    "MarkupSafe==2.1.5",
    "numpy==2.1.1",
    "openpyxl==3.1.5",
    "outcome==1.3.0.post0",
    "packaging==24.1",
    "pandas==2.2.2",
    "pdfkit==1.0.0",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.13.1",
    "PySocks==1.7.1",
    "pytest-playwright>=0.7.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.1",
    "psycopg==3.2.9",
    "selenium==4.24.0",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.6",
    "SQLAlchemy==2.0.35",
    "trio==0.26.2",
    "trio-websocket==0.11.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.1",
    "urllib3==2.2.2",
    "websocket-client==1.8.0",
    "Werkzeug==3.0.4",
    "wsproto==1.2.0",
    "psycopg2-binary>=2.9.11",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.uv]
package = false

